viewmodel implementation and fixing

+++ app/src/main/java/com/example/projecte_uf1/Dimensions.java
@@ -2,27 +2,27 @@ package com.example.projecte_uf1;
 
 public class Dimensions {
 
-    private int width;
-    private int height;
+    private float width;
+    private float height;
 
-    public Dimensions(int width, int height) {
+    public Dimensions(float width, float height) {
         this.width = width;
         this.height = height;
     }
 
-    public int getWidth() {
+    public float getWidth() {
         return width;
     }
 
-    public void setWidth(int width) {
+    public void setWidth(float width) {
         this.width = width;
     }
 
-    public int getHeight() {
+    public float getHeight() {
         return height;
     }
 
-    public void setHeight(int height) {
+    public void setHeight(float height) {
         this.height = height;
     }
 }
